added last admin methods, added site icon to topbar

+++ games4you/src/main/java/games4you/entities/Admin.java
@@ -81,7 +81,7 @@ public class Admin extends User {
         }
     }
 
-    public boolean evaluateReportedReview(int rid, boolean judgment) {
+    public boolean evaluateReportedReview(long rid, boolean judgment) {
         if(!judgment) {
             return super.removeReview(rid);
         }
